+++ GLView/Core/GLUtils.h
@@ -2,15 +2,14 @@
 //  GLUtils.h
 //
 //  GLView Project
-//  Version 1.3.9
+//  Version 1.4
 //
 //  Created by Nick Lockwood on 04/06/2012.
 //  Copyright 2011 Charcoal Design
 //
 //  Distributed under the permissive zlib License
-//  Get the latest version from either of these locations:
+//  Get the latest version from here:
 //
-//  http://charcoaldesign.co.uk/source/cocoa#glview
 //  https://github.com/nicklockwood/GLView
 //
 //  This software is provided 'as-is', without any express or implied
@@ -35,7 +34,7 @@
 //
 //  ARC Helper
 //
-//  Version 1.3.1
+//  Version 2.1
 //
 //  Created by Nick Lockwood on 05/01/2012.
 //  Copyright 2012 Charcoal Design
@@ -46,21 +45,16 @@
 //  https://gist.github.com/1563325
 //
 
-#ifndef AH_RETAIN
+#ifndef ah_retain
 #if __has_feature(objc_arc)
-#define AH_RETAIN(x) (x)
-#define AH_RELEASE(x) (void)(x)
-#define AH_AUTORELEASE(x) (x)
-#define AH_SUPER_DEALLOC (void)(0)
-#define __AH_BRIDGE __bridge
+#define ah_retain self
+#define ah_dealloc self
+#define release self
+#define autorelease self
 #else
-#define __AH_WEAK
-#define AH_WEAK assign
-#define AH_RETAIN(x) [(x) retain]
-#define AH_RELEASE(x) [(x) release]
-#define AH_AUTORELEASE(x) [(x) autorelease]
-#define AH_SUPER_DEALLOC [super dealloc]
-#define __AH_BRIDGE
+#define ah_retain retain
+#define ah_dealloc dealloc
+#define __bridge
 #endif
 #endif
 
@@ -78,22 +72,7 @@ void CGRectGetGLCoords(CGRect rect, GLfloat *coords);
 
 @interface NSString (GL)
 
-- (NSString *)stringByAppendingScaleSuffix;
-- (NSString *)stringByDeletingScaleSuffix;
-- (NSString *)scaleSuffix;
-- (CGFloat)scale;
-
-- (NSString *)stringByAppendingInterfaceIdiomSuffix;
-- (NSString *)stringByDeletingInterfaceIdiomSuffix;
-- (NSString *)interfaceIdiomSuffix;
-- (UIUserInterfaceIdiom)interfaceIdiom;
-
-- (NSString *)stringByAppendingHDSuffix;
-- (NSString *)stringByDeletingHDSuffix;
-- (NSString *)HDSuffix;
-- (BOOL)isHD;
-
-- (NSString *)absolutePathWithDefaultExtensions:(NSString *)firstExtension, ... NS_REQUIRES_NIL_TERMINATION;
+- (NSString *)normalizedPathWithDefaultExtension:(NSString *)extension;
 
 @end