+++ GLView/Core/GLUtils.m
@@ -2,15 +2,14 @@
 //  GLUtils.m
 //
 //  GLView Project
-//  Version 1.3.9
+//  Version 1.4
 //
 //  Created by Nick Lockwood on 04/06/2012.
 //  Copyright 2011 Charcoal Design
 //
 //  Distributed under the permissive zlib License
-//  Get the latest version from either of these locations:
+//  Get the latest version from here:
 //
-//  http://charcoaldesign.co.uk/source/cocoa#glview
 //  https://github.com/nicklockwood/GLView
 //
 //  This software is provided 'as-is', without any express or implied
@@ -33,6 +32,7 @@
 //
 
 #import "GLUtils.h"
+#import <objc/message.h>
 
 
 void CGRectGetGLCoords(CGRect rect, GLfloat *coords)
@@ -50,267 +50,55 @@ void CGRectGetGLCoords(CGRect rect, GLfloat *coords)
 
 @implementation NSString (GL)
 
-- (NSString *)stringByAppendingScaleSuffix
+- (NSString *)normalizedPathWithDefaultExtension:(NSString *)extension
 {
-    if ([UIScreen mainScreen].scale > 1.0f)
-    {
-        NSString *extension = [self pathExtension];
-        NSString *deviceSuffix = [self interfaceIdiomSuffix];
-        NSString *scaleSuffix = [NSString stringWithFormat:@"@%ix", (int)[UIScreen mainScreen].scale];
-        NSString *path = [[self stringByDeletingPathExtension] stringByDeletingInterfaceIdiomSuffix];
-        path = [path stringByAppendingFormat:@"%@%@", scaleSuffix, deviceSuffix];
-        return [extension length]? [path stringByAppendingPathExtension:extension]: path;
-    }
-    return self;
-}
-
-- (NSString *)stringByDeletingScaleSuffix
-{
-    NSString *scaleSuffix = [self scaleSuffix];
-    if ([scaleSuffix length])
-    {
-        NSString *extension = [self pathExtension];
-        NSString *deviceSuffix = [self interfaceIdiomSuffix];
-        NSString *path = [[self stringByDeletingPathExtension] stringByDeletingInterfaceIdiomSuffix];
-        path = [[path substringToIndex:[path length] - [scaleSuffix length]] stringByAppendingString:deviceSuffix];
-        return [extension length]? [path stringByAppendingPathExtension:extension]: path;
-    }
-    return self;
-}
-
-- (NSString *)scaleSuffix
-{
-    //note: this isn't very robust as it only handles single-digit integer scales
-    //for the forseeable future though, it's unlikely that we'll have to worry about that
-    NSString *path = [[self stringByDeletingPathExtension] stringByDeletingInterfaceIdiomSuffix];
-    if ([path length] >= 3)
-    {
-        NSString *scaleSuffix = [path substringFromIndex:[path length] - 3];
-        if ([[scaleSuffix substringToIndex:1] isEqualToString:@"@"] &&
-            [[scaleSuffix substringFromIndex:2] isEqualToString:@"x"])
-        {
-            return scaleSuffix;
-        }
-    }
-    return @"";
-}
-
-- (CGFloat)scale
-{
-    NSString *scaleSuffix = [self scaleSuffix];
-    if ([scaleSuffix length])
-    {
-        return [[scaleSuffix substringWithRange:NSMakeRange(1, 1)] floatValue];
-    }
-    else if (UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPhone && [self isHD])
-    {
-        return 2.0f;
-    }
-    return 1.0f;
-}
-
-- (NSString *)stringByAppendingInterfaceIdiomSuffix
-{
-    NSString *extension = [self pathExtension];
-    NSString *suffix = (UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad)? @"~ipad": @"~iphone";
-    NSString *path = [[self stringByDeletingPathExtension] stringByAppendingString:suffix];
-    return [extension length]? [path stringByAppendingPathExtension:extension]: path;
-}
-
-- (NSString *)stringByDeletingInterfaceIdiomSuffix
-{
-    NSString *suffix = [self interfaceIdiomSuffix];
-    if ([suffix length])
-    {
-        NSString *extension = [self pathExtension];
-        NSString *path = [self stringByDeletingPathExtension];
-        path = [path substringToIndex:[path length] - [suffix length]];
-        return [extension length]? [path stringByAppendingPathExtension:extension]: path;
-    }
-    return self;
-}
-
-- (NSString *)interfaceIdiomSuffix
-{
-    NSString *path = [self stringByDeletingPathExtension];
-    if ([path hasSuffix:@"~iphone"])
-    {
-        return @"~iphone";
-    }
-    else if ([path hasSuffix:@"~ipad"])
-    {
-        return @"~ipad";
-    }
-    return @"";
-}
-
-- (UIUserInterfaceIdiom)interfaceIdiom
-{
-    if ([[self interfaceIdiomSuffix] isEqualToString:@"~ipad"])
-    {
-        return UIUserInterfaceIdiomPad;
-    }
-    else if ([[self interfaceIdiomSuffix] isEqualToString:@"~iphone"])
-    {
-        return UIUserInterfaceIdiomPhone;
-    }
-    return UI_USER_INTERFACE_IDIOM();
-}
-
-- (NSString *)stringByAppendingHDSuffix
-{
-    if ([UIScreen mainScreen].scale > 1.0f || UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad)
-    {
-        NSString *extension = [self pathExtension];
-        NSString *deviceSuffix = [self interfaceIdiomSuffix];
-        NSString *scaleSuffix = [self scaleSuffix];
-        NSString *path = [[[self stringByDeletingPathExtension] stringByDeletingInterfaceIdiomSuffix] stringByDeletingScaleSuffix];
-        path = [path stringByAppendingFormat:@"-hd%@%@", scaleSuffix, deviceSuffix];
-        return [extension length]? [path stringByAppendingPathExtension:extension]: path;
-    }
-    return self;
-}
-
-- (NSString *)stringByDeletingHDSuffix
-{
-    NSString *HDSuffix = [self HDSuffix];
-    if ([HDSuffix length])
-    {
-        NSString *extension = [self pathExtension];
-        NSString *deviceSuffix = [self interfaceIdiomSuffix];
-        NSString *scaleSuffix = [self scaleSuffix];
-        NSString *path = [[[self stringByDeletingPathExtension] stringByDeletingInterfaceIdiomSuffix] stringByDeletingScaleSuffix];
-        path = [path substringToIndex:[path length] - [HDSuffix length]];
-        path = [path stringByAppendingFormat:@"%@%@", scaleSuffix, deviceSuffix];
-        return [extension length]? [path stringByAppendingPathExtension:extension]: path;
-    }
-    return self;
-}
-
-- (NSString *)HDSuffix
-{
-    NSString *path = [[[self stringByDeletingPathExtension] stringByDeletingInterfaceIdiomSuffix] stringByDeletingScaleSuffix];
-    if ([path hasSuffix:@"-hd"])
-    {
-        return @"-hd";
-    }
-    return @"";
-}
-
-- (BOOL)isHD
-{
-    return [[self HDSuffix] length] > 0;
-}
-
-- (NSString *)absolutePathWithDefaultExtensions:(NSString *)firstExtension, ...
-{
-    //set up cache
-    static NSCache *cache = nil;
-    if (cache == nil)
-    {
-        cache = [[NSCache alloc] init];
-    }
-    
-    //convert to absolute path
+    //extension
     NSString *path = self;
-    if (![path isAbsolutePath])
+    if (![[self pathExtension] length])
     {
-        path = [[[NSBundle mainBundle] resourcePath] stringByAppendingPathComponent:path];
-    }
-
-    //add file extension if not already set
-    NSString *extension = [path pathExtension];
-    if ([extension length] == 0)
-    {
-        va_list arguments;
-        va_start(arguments, firstExtension);
-        extension = firstExtension;
-        while (extension)
-        {
-            NSString *_path = [path stringByAppendingPathExtension:extension];
-            if ([[NSFileManager defaultManager] fileExistsAtPath:_path])
-            {
-                path = _path;
-                break;
-            }
-            extension = va_arg(arguments, NSString *);
-        }
-        va_end(arguments);
-    }
-    
-    //check cache
-    NSString *cacheKey = path;
-    BOOL cachable = [path hasPrefix:[[NSBundle mainBundle] resourcePath]];
-    if (cachable)
-    {
-        NSString *_path = [cache objectForKey:cacheKey];
-        if (_path)
-        {
-            return [_path length]? _path: nil;
-        }
-    }
-    
-    if (UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad)
-    {
-        //check for HD version
-        NSString *_path = [path stringByAppendingHDSuffix];
-        if ([[NSFileManager defaultManager] fileExistsAtPath:_path])
-        {
-            path = _path;
-        }
-        
-        //check for scaled version
-        if ([UIScreen mainScreen].scale > 1.0f && [path scale] == 1.0f)
-        {
-            NSString *_path = [path stringByAppendingScaleSuffix];
-            if ([[NSFileManager defaultManager] fileExistsAtPath:_path])
-            {
-                path = _path;
-            }
-        }
-    }
-    else if ([UIScreen mainScreen].scale > 1.0f && [path scale] == 1.0f)
-    {
-        //check for HD version
-        NSString *_path = [path stringByAppendingHDSuffix];
-        if ([[NSFileManager defaultManager] fileExistsAtPath:_path])
-        {
-            path = _path;
-        }
-        else
-        {
-            //check for scaled version
-            NSString *_path = [path stringByAppendingScaleSuffix];
-            if ([[NSFileManager defaultManager] fileExistsAtPath:_path])
-            {
-                path = _path;
-            }
-        }
-    }
-    
-    if ([[NSFileManager defaultManager] fileExistsAtPath:path])
-    {
-        //check for ipad/iphone version
-        NSString *_path = [path stringByAppendingInterfaceIdiomSuffix];
-        if ([[NSFileManager defaultManager] fileExistsAtPath:_path])
-        {
-            path = _path;
-        }
+        path = [path stringByAppendingPathExtension:extension];
     }
     else
     {
-        //file doesn't exist
-        path = nil;
+        extension = [path pathExtension];
     }
     
-    //add to cache
-    if (cachable)
+    //use StandardPaths if available
+    SEL normalizedPathSelector = NSSelectorFromString(@"normalizedPathForFile:");
+    NSFileManager *fileManager = [NSFileManager defaultManager];
+    if ([fileManager respondsToSelector:normalizedPathSelector])
     {
-        [cache setObject:path ?: @"" forKey:cacheKey];
+        return objc_msgSend(fileManager, normalizedPathSelector, path);
+    }
+
+    //convert to absolute path
+    if (![self isAbsolutePath])
+    {
+        path = [[[NSBundle mainBundle] resourcePath] stringByAppendingPathComponent:path];
     }
     
-    //return path
-    return path;
+    //check for @2x version
+    if ([UIScreen mainScreen].scale == 2.0f)
+    {
+        NSString *retinaPath = [[[path stringByDeletingPathExtension] stringByAppendingString:@"@2x"] stringByAppendingPathExtension:extension];
+        if ([fileManager fileExistsAtPath:retinaPath])
+        {
+            path = retinaPath;
+        }
+    }
+    
+    //check for ~ipad version
+    if (UI_USER_INTERFACE_IDIOM() == UIUserInterfaceIdiomPad)
+    {
+        NSString *iPadPath = [[[path stringByDeletingPathExtension] stringByAppendingString:@"~ipad"] stringByAppendingPathExtension:extension];
+        if ([fileManager fileExistsAtPath:iPadPath])
+        {
+            path = iPadPath;
+        }
+    }
+    
+    //default path
+    return [fileManager fileExistsAtPath:path]? path: nil;
 }
 
 @end