+++ GLView/Models/GLLight.m
@@ -2,15 +2,14 @@
 //  GLLight.m
 //
 //  GLView Project
-//  Version 1.3.9
+//  Version 1.4
 //
 //  Created by Nick Lockwood on 17/05/2012.
 //  Copyright 2011 Charcoal Design
 //
 //  Distributed under the permissive zlib License
-//  Get the latest version from either of these locations:
+//  Get the latest version from here:
 //
-//  http://charcoaldesign.co.uk/source/cocoa#glview
 //  https://github.com/nicklockwood/GLView
 //
 //  This software is provided 'as-is', without any express or implied
@@ -77,10 +76,10 @@
 
 - (void)dealloc
 {
-    AH_RELEASE(_ambientColor);
-    AH_RELEASE(_diffuseColor);
-    AH_RELEASE(_specularColor);
-    AH_SUPER_DEALLOC;
+    [_ambientColor release];
+    [_diffuseColor release];
+    [_specularColor release];
+    [super ah_dealloc];
 }
 
 @end