+++ GLView/Models/GLModel.h
@@ -2,15 +2,14 @@
 //  GLModel.h
 //
 //  GLView Project
-//  Version 1.3.9
+//  Version 1.4
 //
 //  Created by Nick Lockwood on 10/07/2011.
 //  Copyright 2011 Charcoal Design
 //
 //  Distributed under the permissive zlib License
-//  Get the latest version from either of these locations:
+//  Get the latest version from here:
 //
-//  http://charcoaldesign.co.uk/source/cocoa#glview
 //  https://github.com/nicklockwood/GLView
 //
 //  This software is provided 'as-is', without any express or implied
@@ -39,10 +38,11 @@
 
 @interface GLModel : NSObject
 
-+ (GLModel *)modelWithContentsOfFile:(NSString *)path;
++ (GLModel *)modelNamed:(NSString *)nameOrPath;
++ (GLModel *)modelWithContentsOfFile:(NSString *)nameOrPath;
 + (GLModel *)modelWithData:(NSData *)data;
 
-- (GLModel *)initWithContentsOfFile:(NSString *)path;
+- (GLModel *)initWithContentsOfFile:(NSString *)nameOrPath;
 - (GLModel *)initWithData:(NSData *)data;
 
 - (void)draw;